requirements.txt
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0
scikit-image>=0.21.0 